+++ src/controller/index.ts
@@ -1,0 +1,1 @@
+export {readCode} from './oauth/github';
