+++ src/route/index.ts
@@ -1,0 +1,1 @@
+export {githubOauth} from './githubOauth';
